03_query_ai/hw1/requirements.txt
streamlit
pandas
plotly
python-dotenv
requests
massive
